Finding mate upgrade - one more try

+++ src/Figures/Bishop.java
@@ -3,12 +3,12 @@ package Figures;
 import Board.Board;
 
 public class Bishop extends Figure{
-    public Bishop(boolean isWhite, int x, int y) {
+    public Bishop(boolean isWhite, byte x, byte y) {
         super(isWhite, x, y, (byte) 4);
     }
 
     @Override
-    public boolean isGoodToGo(Board board, int x, int y) {
+    public boolean isGoodToGo(Board board, byte x, byte y) {
         if(board.getSquareBoard()[y][x].isOccupied() && board.getSquareBoard()[y][x].getFigure().isWhite() == this.isWhite){
             return false;
         }
