+++ src/Figures/King.java
@@ -4,12 +4,12 @@ package Figures;
 import Board.Board;
 
 public class King extends Figure{
-    public King(boolean isWhite, int x, int y) {
+    public King(boolean isWhite, byte x, byte y) {
         super(isWhite, x, y, (byte) 6);
     }
 
     @Override
-    public boolean isGoodToGo(Board board, int x, int y) {
+    public boolean isGoodToGo(Board board, byte x, byte y) {
         if(board.getSquareBoard()[y][x].isOccupied() && board.getSquareBoard()[y][x].getFigure().isWhite() == this.isWhite){
             return false;
         }
