+++ src/Figures/Queen.java
@@ -5,13 +5,13 @@ import Figures.Rook;
 import Figures.Bishop;
 
 public class Queen extends Figure {
-    public Queen(boolean isWhite, int x, int y) {
+    public Queen(boolean isWhite, byte x, byte y) {
         super(isWhite, x, y, (byte) 5);
     }
 
 
     @Override
-    public boolean isGoodToGo(Board board, int x, int y) {
+    public boolean isGoodToGo(Board board, byte x, byte y) {
         if(board.getSquareBoard()[y][x].isOccupied() && board.getSquareBoard()[y][x].getFigure().isWhite() == this.isWhite){
             return false;
         }
